requirements.txt

imageio==2.13.5
numpy==1.21.6
opencv-python==4.5.5.62
opt-einsum==3.3.0
pandas==1.3.5
Pillow==8.4.0
pytorch-msssim==0.2.1
pytorch-warmup==0.1.0
pytorch-wavelets @ file:///media2/inder/ruchit_mount_2/pytorch_wavelets
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
torch==1.13.0.dev20220624+cu113
torch-lr-finder==0.2.1
torch-tb-profiler==0.4.0
torchattacks==3.2.4
torchmetrics==0.10.2
torchvision==0.14.0.dev20220624+cu113
tqdm==4.62.3
wandb==0.13.7
